Review SSAO

+++ common/src/tools/svg/shape/SVGCircle.h
@@ -14,7 +14,7 @@ namespace urchin
             SVGCircle(const Point2<float> &, float, SVGColor, float opacity = 1.0f);
 
             std::string getShapeTag() const override;
-            Rectangle<int> computeRectangle() const override;
+            Rectangle<float> computeRectangle() const override;
 
         private:
             Point2<float> center;
