+++ common/src/tools/svg/shape/SVGShape.h
@@ -25,7 +25,7 @@ namespace urchin
             void setStroke(SVGColor, float);
 
             virtual std::string getShapeTag() const = 0;
-            virtual Rectangle<int> computeRectangle() const = 0;
+            virtual Rectangle<float> computeRectangle() const = 0;
 
         protected:
             std::string getStyle() const;
